Update _PS_Library_Groups.psm1
As part of the New Starter stuff, have created a new Term Group for Live Teams (based on Sharepoint Team Sites) - addto-SharepointTeamsTermStore in this script will check $displayName of newly created team site for "Sym" or "Working Group", if they don't match it will add the $displayName to the Term Store. This will allow People Services to pick live teams (currently it shows all groups via the Sharpeoint 365 People and Group Picker, so the wrong names go in - such as the Owners Group). Hoping this will improve some automation for Data Managers and be a bit clearer for users for this form.